refactor(ors_control_app): extract routes/query and routes/static
Move SPA serving, tile proxy, query passthrough, and logout redirect into
focused Router modules:
- routes/query.ts: /api/query, /api/tiles/:z/:x/:y, /logout
- routes/static.ts: /assets, express.static, SPA fallback

server/index.ts mounts via app.use(createQueryRouter()) and
app.use(createStaticRouter(distDir)). Lines down 3,085 -> 3,001.

Phase 1 of holistic-control-app-refactor-final, route extraction step 3.
No behaviour change. tsc passes cleanly.
